+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/dto/MaintenanceCalculatedDataDTO.java
@@ -1,12 +1,11 @@
 package com.wks.caseengine.dto;
 
-import lombok.*;
-
-import java.util.UUID;
-
-import org.springframework.context.annotation.Configuration;
-
 import com.wks.caseengine.entity.MaintenanceCalculatedData;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 //@Configuration
 @NoArgsConstructor
 @AllArgsConstructor
@@ -15,12 +14,12 @@ import com.wks.caseengine.entity.MaintenanceCalculatedData;
 public class MaintenanceCalculatedDataDTO {
     private String id;
     private Integer runningHoursInMonth;
-    private Float shoutdownHrs;
+    private Double shoutdownHrs;
     private Integer nonShoutdownHrs;
-    private Float eoeAvgSlowdownLoadPVT;
-    private Float eoAvgSlowdownLoadPVT; 
-    private Float eoeSlowdownLoadReduction;
-    private Float eoSlowdownLoadReduction;
+    private Double eoeAvgSlowdownLoadPVT;
+    private Double eoAvgSlowdownLoadPVT; 
+    private Double eoeSlowdownLoadReduction;
+    private Double eoSlowdownLoadReduction;
     private Integer eoeEffectiveOperatingHrs;
     private Integer eoEffectiveOperatingHrs;
     private Integer monthNo;
